plan: lower MaxDecompressedPushPlan to 64 MiB and stop ReadAll's doubling growth (task 3g2)
be2's own 256 MiB cap was 4x the "tens of MB" legitimate envelope its own
doc comment described, and io.ReadAll's internal doubling-growth buffer
meant even a stream entirely INSIDE that cap (no refusal at all) could
still peak at roughly 2x the cap in memory. Measured pre-fix: a 254,845
compressed-byte input decompressing to 250 MiB peaked at ~924 MB RSS
(~3,630x), and a 203,883-byte input decompressing to 200 MiB peaked at
~956 MB RSS (~4,700x) — both fully policy-compliant under the old cap, so
nothing refused either one, on a fleet that includes 1 GB-class SBCs.

Two independent changes:
- MaxDecompressedPushPlan is now a var defaulting to 64 MiB (still >100x
  every real recorded plan checked) instead of a 256 MiB const, so an
  embedder with a genuinely large legitimate plan can raise it instead of
  hitting a hard wall.
- maybeGunzip's io.ReadAll(io.LimitReader(gr, max+1)) is replaced by
  readCapped, which reads in small fixed-size chunks into a bytes.Buffer
  pre-grown to just one chunk and refuses the moment the running byte
  count would exceed max, before ever writing the crossing bytes into the
  buffer — so a stream well past the cap never grows the buffer anywhere
  near it, let alone past it.

Post-fix, the same two probe sizes peak at ~139 MB RSS (both refused
immediately at the new lower cap), reproduced via revert-and-retest
against the pre-fix code. Real recorded plans from this repo's own tests,
dotfiles/gonf and conf/gonf all round-trip through the exact
EncodePush/DecodePush path with 100x-3000x headroom under the new cap, and
TestCLIApplySealedLargeLegitimatePlanSucceeds confirms a real ~7.8 MiB
sealed apply still succeeds end-to-end.

+++ internal/cli/cli_sealed_apply_test.go
@@ -950,7 +950,7 @@ func TestCLIApplySealedLargeLegitimatePlanSucceeds(t *testing.T) {
 	frame := buf.Bytes()
 
 	sealed := sealFrame(t, frame, []string{recipientLine})
-	if len(sealed) > maxSealedFrameBytes || len(sealed) > plan.MaxDecompressedPushPlan {
+	if len(sealed) > maxSealedFrameBytes || int64(len(sealed)) > plan.MaxDecompressedPushPlan {
 		t.Fatalf("test construction error: sealed frame %d bytes already exceeds a cap, not a meaningful legitimate-plan check", len(sealed))
 	}
 	planPath := writeSealedPlanFile(t, dir, sealed)
